refactor(links): handle selection changes in plugin
